Pattern matching to determine volume number
Use a regular expression to match the volume number as a digit. The pattern matches the word 'volume', 'vol', '#', or 'v', followed by one or more digits. The digits are captured in a group that can be accessed later.

This might need some work later. I'm not sure that all of these are necessary.